Create Chunk <-> Json bijection that correctly handles multiple chunks.
Shockingly, the unapply method for BijectionsChunkJson was incorrect, only reading the first chunk of data, and not following the link to the next chunk. This is a problem if using the client and receiving a large amount of data. This add a custom implementation that fixes the issue, and migrates Bigtop utilities to use the fixed bijection.

We also correctly encode output as utf-8, instead of relying on the default charset.

Will merge in BE 0.6 in due course.